remove formatgen
This was more painful than I originally thought, and the possibility for
it to be useful was not outweighing its pain-points. There is a lot of
one off stuff that was not possible to implement in a IntoCoder, so it
meant that functionality was pushed onto Reader, where it doesn't
belong.